WIP :added pager for device list
