+++ clearblade/cloud/iot_v1/client.py
@@ -1,5 +1,6 @@
 from devices import *
 from registry import *
+from device_types import *
 
 class DeviceManagerClient():
 
@@ -41,15 +42,15 @@ class DeviceManagerClient():
     def list_device_states(self, request : GetDeviceStatesList):
         cb_device_manager = ClearBladeDeviceManager()
         return cb_device_manager.getDeviceSatesList(request=request)
-    
+
     def list_device_config_versions(self, request : GetDeviceConfigVersionsList):
         cb_device_manager = ClearBladeDeviceManager()
         return cb_device_manager.getDeviceConfigVersionsList(request=request)
-    
+
     def list_devices(self, request : ListDevicesRequest):
         cb_device_manager = ClearBladeDeviceManager()
-        return cb_device_manager.getDevicesList(request=request)
-    
+        return cb_device_manager.list(request=request)
+
     def update_device(self, request : UpdateDeviceRequest):
         cb_device_manager = ClearBladeDeviceManager()
         return cb_device_manager.updateDevice(request=request)
@@ -99,14 +100,14 @@ class DeviceManagerAsyncClient():
     async def list_device_states(self, request : GetDeviceStatesList):
         cb_device_manager = ClearBladeDeviceManager()
         return await cb_device_manager.getDeviceSatesList_async(request=request)
-    
+
     async def list_device_config_versions(self, request : GetDeviceConfigVersionsList):
         cb_device_manager = ClearBladeDeviceManager()
         return await cb_device_manager.getDeviceConfigVersionsList_async(request=request)
 
     async def list_devices(self, request : ListDevicesRequest):
         cb_device_manager = ClearBladeDeviceManager()
-        return await cb_device_manager.getDevicesList_async(request=request)
+        return await cb_device_manager.list_async(request=request)
 
     async def update_device(self, request : UpdateDeviceRequest):
         cb_device_manager = ClearBladeDeviceManager()
