+++ clearblade/cloud/iot_v1/devices.py
@@ -1,325 +1,7 @@
 from http_client import SyncClient, AsyncClient
-from config import ClearBladeConfig
 from config_manager import ClearBladeConfigManager
-
-class Device():
-    """
-    Data class for Clearblade Device
-    """
-    #TODO: find a better way to construct the Device object. I dont like so much parameter in a constructor
-    def __init__(self, id: str = None, name: str = None, num_id: str=None,
-                 credentials: list=[], last_heartbeat_time: str = None, last_event_time: str = None,
-                 last_state_time: str = None, last_config_ack_time: str = None,
-                 last_config_send_time: str = None, blocked: bool = False,
-                 last_error_time: str = None, last_error_status_code: dict = {"code":None, "message":""},
-                 config: dict = {"cloudUpdateTime":None, "version":""} ,
-                 state: dict = {"updateTime":None, "binaryData":None},
-                 log_level: str = "NONE", meta_data: dict = {}, gateway_config : dict = {}  ) -> None:
-
-        self._id = id
-        self._name = name
-        self._num_id = num_id
-        self._credentials = credentials
-        self._last_heartbeat_time = last_heartbeat_time
-        self._last_event_time = last_event_time
-        self._last_state_time = last_state_time
-        self._last_config_ack_time = last_config_ack_time
-        self._last_config_send_time = last_config_send_time
-        self._blocked = blocked
-        self._last_error_time = last_error_time
-        self._last_error_status_code = last_error_status_code
-        self._config = config
-        self._state =  state
-        self._log_level = log_level
-        self._meta_data = meta_data
-        self._gateway_config = gateway_config
-
-    @staticmethod
-    def from_json(json):
-        return Device(id= json['id'], name= json['name'], num_id= json['numId'],
-                      credentials= json['credentials'], last_heartbeat_time= json['lastHeartbeatTime'],
-                      last_event_time= json['lastEventTime'], last_state_time= json['lastStateTime'],
-                      last_config_ack_time= json['lastConfigAckTime'], last_config_send_time= json['lastConfigSendTime'],
-                      blocked= json['blocked'], last_error_time= json['lastErrorTime'],
-                      last_error_status_code= json['lastErrorStatus'], config= json['config'],
-                      state= json['state'], log_level= json['logLevel'], meta_data= json['metadata'],
-                      gateway_config= json['gatewayConfig'])
-
-    @property
-    def id(self):
-        return self._id
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def num_id(self):
-        return self._num_id
-
-    @property
-    def credentials(self):
-        return self._credentials
-
-    @property
-    def last_error_status(self):
-        return self._last_error_status_code
-
-    @property
-    def config(self):
-        return self._config
-
-    @property
-    def state(self):
-        return self._state
-
-    @property
-    def log_level(self):
-        return self._log_level
-
-    @property
-    def meta_data(self):
-        return self._meta_data
-
-    @property
-    def gateway_config(self):
-        return self._gateway_config
-
-    @property
-    def log_level(self):
-        return self._log_level
-
-#classes to mock googles request & response
-
-class Request():
-    def __init__(self, name) -> None:
-        self._name = name
-
-    @property
-    def name(self):
-        return self._name
-
-class SendCommandToDeviceRequest(Request):
-    def __init__(self, name: str = None,
-                binary_data: bytes = None,
-                subfolder: str = None) -> None:
-        super().__init__(name)
-        self._binary_data = binary_data
-        self._subfolder = subfolder
-
-    @property
-    def binary_data(self):
-        return self._binary_data
-
-    @property
-    def sub_folder(self):
-        return self._subfolder
-
-class CreateDeviceRequest(Request):
-    def __init__(self, name: str = None,
-                       device: Device = None) -> None:
-        super().__init__(name)
-        self._device = device
-
-    @property
-    def device(self):
-        return self._device
-
-class ModifyCloudToDeviceConfigRequest(Request):
-    def __init__(self, name:str = None ,
-                 version_to_update: int = -1,
-                 binary_data: bytes = None) -> None:
-        super().__init__(name)
-        self._version_to_update = version_to_update
-        self._binary_data = binary_data
-
-    @property
-    def version_to_update(self):
-        return self._version_to_update
-
-    @property
-    def binary_data(self):
-        return self._binary_data
-
-class DeviceConfig(Request):
-    def __init__(self, name,
-                 version,
-                 cloud_ack_time,
-                 device_ack_time,
-                 binary_data) -> None:
-        super().__init__(name)
-        self._version = version
-        self._cloud_ack_time = cloud_ack_time
-        self._device_ack_time = device_ack_time
-        self._binary_data = binary_data
-
-    def version(self):
-        return self._version
-
-    def cloud_ack_time(self):
-        return self._cloud_ack_time
-
-    def device_ack_time(self):
-        return self._device_ack_time
-
-    def binary_data(self):
-        return self._binary_data
-
-    @staticmethod
-    def from_json(json):
-        return DeviceConfig(version=json['version'], cloud_ack_time=json['cloudUpdateTime'],
-                            device_ack_time=json['deviceAckTime'], binary_data=json['binaryData'])
-
-class DeleteDeviceRequest(Request):
-    def __init__(self, name: str = None) -> None:
-        super().__init__(name)
-
-class GetDeviceRequest(Request):
-    def __init__(self, name: str = None) -> None:
-        super().__init__(name)
-
-class BindUnBindGatewayDeviceRequest(Request):
-    def __init__(self, deviceId: str = None,
-                       gatewayId: str = None) -> None:
-        self._deviceid=deviceId
-        self._gatewayid=gatewayId
-
-    @property
-    def deviceId(self):
-        return self._deviceid
-
-    @property
-    def gatewayId(self):
-        return self._gatewayid
-
-class GetDeviceStatesList(Request):
-    def __init__(self, name: str = None,
-                numStates: int = None) -> None:
-        super().__init__(name)
-        self._numstates = numStates
-
-    @property
-    def numStates(self):
-        return self._numstates
-
-class GetDeviceConfigVersionsList(Request):
-    def __init__(self, name: str = None,
-                numVersions: int = None) -> None:
-        super().__init__(name)
-        self._numversions = numVersions
-
-    @property
-    def numVersions(self):
-        return self._numversions
-
-class ListDevicesRequest(Request):
-    def __init__(self, parent:str = None ,
-                 deviceNumIds: str = None,
-                 deviceIds: str = None,
-                 fieldMask: str = None,
-                 gatewayListOptions: dict = None,
-                 pageSize: int = None,
-                 pageToken: str = None) -> None:
-        self._parent = parent
-        self._device_num_ids = deviceNumIds
-        self._device_ids = deviceIds
-        self._field_mask = fieldMask
-        self._gateway_list_options = gatewayListOptions
-        self._page_size = pageSize
-        self._page_token = pageToken
-
-    @property
-    def parent(self):
-        return self._parent
-
-    @property
-    def device_num_ids(self):
-        return self._device_num_ids
-
-    @property
-    def device_ids(self):
-        return self._device_ids
-
-    @property
-    def field_mask(self):
-        return self._field_mask
-
-    @property
-    def gateway_list_options(self):
-        return self._gateway_list_options
-
-    @property
-    def page_size(self):
-        return self._page_size
-
-    @property
-    def page_token(self):
-        return self._page_token
-
-    def _prepare_params_for_list(self):
-        params = {'parent':self.parent}
-        if self.page_size is not None:
-            params['pageSize'] = self.page_size
-        if self.device_num_ids is not None:
-            params['deviceNumIds'] = self.device_num_ids
-        if self.device_ids is not None:
-            params['deviceIds'] = self.device_ids
-        if self.field_mask is not None:
-            params['fieldMask'] = self.field_mask
-        if self.gateway_list_options is not None:
-            params['gatewayListOptions'] = self.gateway_list_options
-        if self.page_token is not None:
-            params['pageToken'] = self.page_token
-
-class UpdateDeviceRequest(Request):
-    def __init__(self, id: str = None, name: str = None, numId: str=None,
-                 credentials: list=None, blocked: bool = None,
-                 logLevel: str = None, metadata: dict = None, gatewayConfig : dict = None,
-                 updateMask: str = None) -> None:
-        self._id = id
-        self._name = name
-        self._num_id = numId
-        self._credentials = credentials
-        self._blocked = blocked
-        self._log_level = logLevel
-        self._meta_data = metadata
-        self._gateway_config = gatewayConfig
-        self._update_mask = updateMask
-
-    def _prepare_params_body_for_update(self):
-        params = {'name': self._name}
-        if self._update_mask is not None:
-            params['updateMask'] = self._update_mask
-
-        body = {}
-        if self._id is not None:
-            body['id'] = self._id
-        if self._name is not None:
-            body['name'] = self._name
-        if self._log_level is not None:
-            body['logLevel'] = self._log_level
-        if self._gateway_config is not None:
-            body['gatewayConfig'] = self._gateway_config
-        if self._meta_data is not None:
-            body['metadata'] = self._meta_data
-        if self._blocked is not None:
-            body['blocked'] = self._blocked
-        if self._credentials is not None:
-            body['credentials'] = self._credentials
-
-        return params, body
-
-class ListDevicesResponse():
-
-    @property
-    def raw_pages(self):
-        pass
-
-    def devices(self):
-        pass
-
-    def next_page_token(self):
-        pass
+from pagers import ListDevicesPager
+from device_types import *
 
 class ClearBladeDeviceManager():
 
@@ -466,8 +148,38 @@ class ClearBladeDeviceManager():
             return self._create_device_config(response.json)
         return None
 
-    def list(self):
-        pass
+    def get_device_list_response(self, request:ListDevicesRequest):
+        sync_client = SyncClient(clearblade_config= self._config_manager.regional_config)
+        params = request._prepare_params_for_list()
+        response = sync_client.get(api_name="cloudiot_devices",request_params=params)
+
+        print(response.text)
+
+        if response.status_code == 200:
+            return ListDevicesResponse.from_json(response.json())
+        return None
+
+    def list(self,
+             request: ListDevicesRequest):
+
+        device_list_response = self.get_device_list_response(request=request)
+        #devices = device_list_response.devices
+        #yield from devices
+
+        if device_list_response:
+            return ListDevicesPager(method=self.get_device_list_response,
+                                    request=request, response=device_list_response)
+        return None
+
+    async def list_async(self,
+                         request: ListDevicesRequest):
+        async_client = AsyncClient(clearblade_config=self._config_manager.regional_config)
+        params = request._prepare_params_for_list()
+        response = await async_client.list(api_name="cloudiot_devices",request_params=params)
+
+        if response.status_code == 200:
+            return response.json()
+        return None
 
     def get(self,
             request: GetDeviceRequest) -> Device:
@@ -593,26 +305,6 @@ class ClearBladeDeviceManager():
         async_client = AsyncClient()
         params = {'name':request.name, 'numVersions':request.numVersions}
         response = await async_client.get(api_name="cloudiot_devices_configVersions", request_params=params)
-
-        if response.status_code == 200:
-            return response.json()
-        return None
-
-    def getDevicesList(self,
-                       request: ListDevicesRequest):
-        sync_client = SyncClient()
-        params = request._prepare_params_for_list()
-        response = sync_client.get(api_name="cloudiot_devices",request_params=params)
-
-        if response.status_code == 200:
-            return response.json()
-        return None
-
-    async def getDevicesList_async(self,
-            request: ListDevicesRequest):
-        async_client = AsyncClient()
-        params = request._prepare_params_for_list()
-        response = await async_client.list(api_name="cloudiot_devices",request_params=params)
 
         if response.status_code == 200:
             return response.json()
